Camera rotation mostly working, but also not at all

+++ src/lib/timing/Animation.java
@@ -2,19 +2,19 @@ package lib.timing;
 
 public class Animation {
     private final TimingFunction function;
-    private final int duration;
+    private final long duration;
 
     private boolean finished;
     private long startTime, endTime;
 
-    public Animation(TimingFunction function, int duration){
+    public Animation(TimingFunction function, long duration){
         this.function = function;
         this.duration = duration*1000000;
         startTime = System.nanoTime();
         endTime = startTime + this.duration;
     }
 
-    public Animation(int duration) {
+    public Animation(long duration) {
         this(TimingFunction.EASE, duration);
     }
 
@@ -36,7 +36,7 @@ public class Animation {
         return finished;
     }
 
-    public int getDuration() {
+    public long getDuration() {
         return duration;
     }
 }
